Mongo indexes,performance tunning

+++ parser/parser_algorithm.go
@@ -64,7 +64,7 @@ type jaccardIndexComparator struct{}
 
 func (j *jaccardIndexComparator) Compare(mapA map[string]int, mapB map[string]int) float32 {
 	if len(mapA) == 0 || len(mapB) == 0 {
-		return 0
+		return float32(0)
 	}
 
 	//Compute intersection
